requirements.txt
celery==4.0.2
you-get==0.4.652
PyYAML==3.12
requests==2.13.0
pyquery==1.2.17
SQLAlchemy==1.1.6
PyMySQL==0.7.10
statsd==3.2.1
